final fix

+++ lab6.sql
@@ -25,8 +25,8 @@ CREATE TABLE `stations` (
   `id` tinyint(255) unsigned NOT NULL AUTO_INCREMENT,
   `name` varchar(256) COLLATE utf8_unicode_ci NOT NULL,
   `branch_id` varchar(256) COLLATE utf8_unicode_ci NOT NULL,
-  `next` int(11) NOT NULL,
-  `prev` int(11) NOT NULL,
+  `next` tinyint(255) unsigned NOT NULL,
+  `prev` tinyint(255) unsigned NOT NULL,
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
 
@@ -34,7 +34,7 @@ CREATE TABLE `stations` (
 DROP TABLE IF EXISTS `transition`;
 CREATE TABLE `transition` (
   `station_id` tinyint(255) unsigned NOT NULL,
-  `transition_to_id` int(11) NOT NULL,
+  `transition_to_id` tinyint(255) unsigned NOT NULL,
   KEY `station_id` (`station_id`),
   CONSTRAINT `transition_ibfk_1` FOREIGN KEY (`station_id`) REFERENCES `stations` (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
